+++ CHANGELOG.md
@@ -11,6 +11,7 @@ CHANGE LOG
 * Added repository effective branching model API
 * Added repository pipelines config subresources
 * Added repository permissions config APIs
+* Added project permissions config APIs
 * Fixed result pager pagination when using partial response fields
 * Fixed result pager previous-page navigation
 * Fixed source directory listings for specific revisions